+++ client/effect/effect.dart
@@ -14,3 +14,6 @@
 #source("src/Motion.dart");
 #source("src/EasingMotion.dart");
 #source("src/LinearPathMotion.dart");
+#source("src/BuzzEffect.dart");
+#source("src/GlowEffect.dart");
+#source("src/ShakeEffect.dart");